[871] Moves more generic helper functions for the cache to the helper file
