+++ packages/app/src/__tests__/Cache.test.ts
@@ -9,7 +9,7 @@ import {
   GroupBy,
   ServiceData,
 } from '@cloud-carbon-footprint/common'
-import cache, { paginateRequest } from '../Cache'
+import cache from '../Cache'
 import { EstimationRequest } from '../CreateValidRequest'
 import DurationConstructor = moment.unitOfTime.DurationConstructor
 import LocalCacheManager from '../LocalCacheManager'
@@ -597,52 +597,6 @@ describe('Cache', () => {
           },
         ],
       ])
-    })
-  })
-
-  describe('pagination', () => {
-    it('paginates request for skip = 7, limit = 7 and groupBy day', async () => {
-      const rawRequest: EstimationRequest = {
-        startDate: moment.utc('2020-01-01').toDate(),
-        endDate: moment.utc('2020-01-31').toDate(),
-        ignoreCache: false,
-        groupBy: GroupBy.day,
-        limit: 7,
-        skip: 7,
-      }
-
-      const updatedRawRequest: EstimationRequest = {
-        startDate: moment.utc('2020-01-08').toDate(),
-        endDate: moment.utc('2020-01-14').toDate(),
-        ignoreCache: false,
-        groupBy: GroupBy.day,
-        limit: 7,
-        skip: 7,
-      }
-
-      expect(paginateRequest(rawRequest)).toEqual(updatedRawRequest)
-    })
-
-    it('paginates request using endDate at the end of period date range', async () => {
-      const rawRequest: EstimationRequest = {
-        startDate: moment.utc('2020-01-01').toDate(),
-        endDate: moment.utc('2020-01-31').toDate(),
-        ignoreCache: false,
-        groupBy: GroupBy.day,
-        limit: 10,
-        skip: 25,
-      }
-
-      const updatedRawRequest: EstimationRequest = {
-        startDate: moment.utc('2020-01-26').toDate(),
-        endDate: moment.utc('2020-01-31').toDate(),
-        ignoreCache: false,
-        groupBy: GroupBy.day,
-        limit: 10,
-        skip: 25,
-      }
-
-      expect(paginateRequest(rawRequest)).toEqual(updatedRawRequest)
     })
   })
 })
